Add files via upload
tabata.py is the initial tabata timer I built that takes user input via the ```input``` function and prints messages to user to the console.

tabat_oop.py is just some initial classes to make pages to receive input and display to the user. Need to set class attributes to get and pass the information between methods.